backend/requirements.txt
fastapi==0.109.0
uvicorn==0.23.2
python-multipart==0.0.20
pydantic<3

numpy==1.26.4
faiss-cpu==1.7.4
PyPDF2==3.0.1
pdfplumber==0.9.0

torch==2.1.2+cpu
torchvision==0.16.2+cpu
transformers==4.35.2
sentence-transformers==2.2.2
huggingface_hub==0.16.4

langchain==0.1.12
langchain-community==0.0.28
langchain-groq==0.1.0
groq==0.4.2
